# contracts/openapi/yurlib-v1.yaml
openapi: 3.1.0
info:
  title: Yurlib API
  version: 0.1.0
  description: Initial contract for the first local-library vertical slice.
servers:
  - url: /
paths:
...
  /api/v1/library-roots:
    get:
      operationId: listLibraryRoots
      summary: List configured library roots
      responses:
        "200":
          description: Configured roots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryRoot"
    post:
      operationId: createLibraryRoot
      summary: Configure and validate a typed library root
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLibraryRootRequest"
      responses:
        "201":
          description: Root configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryRoot"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /api/v1/library-roots/{rootId}:
    delete:
      operationId: deleteLibraryRoot
      summary: Remove an unused library root
      security:
        - sessionCookie: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/RootId"
      responses:
        "204":
          description: Root removed
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  parameters:
    RootId:
      name: rootId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: RFC 9457 problem response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    CreateLibraryRootRequest:
      type: object
      additionalProperties: false
      required: [name, mountAlias, relativePath, identityToken, mode]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        mountAlias:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,62}$"
        relativePath:
          type: string
          maxLength: 1024
          description: Normalized path relative to the deployment-owned mount alias.
        identityToken:
          type: string
          minLength: 16
          maxLength: 200
          writeOnly: true
        mode:
          type: string
          enum: [READ_ONLY_SOURCE, MANAGED_OUTPUT]
    LibraryRoot:
      type: object
      required: [id, name, mountAlias, relativePath, mode, availability]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        mountAlias:
          type: string
        relativePath:
          type: string
        mode:
          type: string
          enum: [READ_ONLY_SOURCE, MANAGED_OUTPUT]
        availability:
          $ref: "#/components/schemas/RootAvailability"
        lastSuccessfulScanAt:
          type: [string, "null"]
          format: date-time
    RootAvailability:
      type: string
      enum: [UNKNOWN, AVAILABLE, UNAVAILABLE, IDENTITY_MISMATCH]
...
    Problem:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        correlationId:
          type: string
security:
  - sessionCookie: []